backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=5.1.0",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=3.2.0,<4.0.0",
    "fastapi>=0.133.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart==0.0.18",
    "redis>=7.2.0",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]
